refactor(daemon)!: put the pool transport back behind a feature gate
#3079 deleted the `tensor-pool` feature #3152 had just added, made
`dora-tensor-pool` an unconditional dependency, and flipped the Python
wheel's `tensor-pool` to default-on. It also took `daemon/src/lib.rs` from
0 to 47 `unsafe` — against a seam contract that budgeted "<200 lines of
dora, no new unsafe". The README's documented build command,
`cargo build -p dora-daemon --features tensor-pool`, had stopped existing.

Restores the gate:

  - `tensor-pool = ["dep:dora-tensor-pool", "dep:postcard",
    "dep:shared_memory_extended"]`, off by default
  - `#[cfg(feature = "tensor-pool")] mod memory_pool;`
  - the six pool fields on `Daemon` collapse into one `PoolState`, which
    also takes over `MemoryPoolSubscriber` and its `dataflow_live` flag
    (#3194), and `ensure_cross_data_listener` / the subscriber spawn / the
    finish-time cleanup move in with it — so the gate sits at module
    granularity instead of scattered through 1,200 lines
  - `coordinator::resolve_machine` and `CROSS_REGISTER_TIMEOUT` gate too
  - the Python wheel's default reverts to `["tracing", "metrics", "async"]`

A default `cargo tree -p dora-daemon` no longer contains
`dora-tensor-pool`, and the daemon crate's only remaining `unsafe` is the
7 pre-existing occurrences in `shutdown.rs`, `running_dataflow.rs` and
`spawn/prepared.rs` — unchanged from before #3079.

With the feature off, both opaque carriers answer explicitly rather than
hanging: a peer's `ExtensionMessage` is debug-logged and dropped, and a
node's `ExtensionRequest` gets "no extension registered under {ns}".

`binaries/cli` gains a pass-through, since the CLI hosts the in-process
daemon and the feature would otherwise be unreachable from `dora run`.

251 daemon tests with the feature, 230 without.
