+++ binaries/cli/Cargo.toml
@@ -20,6 +20,11 @@ default = ["tracing", "metrics", "redb-backend"]
 tracing = ["dep:dora-tracing"]
 metrics = ["dora-runtime-shared-lib/metrics", "dora-coordinator/metrics"]
 python = ["pyo3"]
+# Pass-through for the daemon's opt-in tensor-pool transport. The CLI hosts the
+# in-process daemon (`dora run`, `dora up`), so without this the feature is
+# unreachable through the `dora` binary. NOT covered by dora's 1.0
+# compatibility guarantees — see libraries/extensions/tensor-pool/README.md.
+tensor-pool = ["dora-daemon/tensor-pool"]
 # Enable extension-module only when building the Python wheel via maturin.
 # This suppresses Python library linking, which is correct for cdylib targets
 # loaded into Python but breaks standalone binary and test builds.
