+++ binaries/daemon/Cargo.toml
@@ -15,13 +15,17 @@ repository.workspace = true
 default = ["tracing", "telemetry"]
 tracing = ["dep:dora-tracing"]
 bench = []
+# Opt-in cross-machine tensor-pool transport. NOT covered by dora's 1.0
+# compatibility guarantees, and the only source of `unsafe` in this crate —
+# see libraries/extensions/tensor-pool/README.md.
+tensor-pool = ["dep:dora-tensor-pool", "dep:postcard", "dep:shared_memory_extended"]
 # telemetry flag enables to trace dora-daemon as well as send ticks with opentelemetry context
 # for distributed tracing.
 telemetry = ["dep:tracing-opentelemetry"]
 
 [dependencies]
-dora-tensor-pool = { workspace = true }
-postcard = { workspace = true }
+dora-tensor-pool = { workspace = true, optional = true }
+postcard = { workspace = true, optional = true }
 eyre = { workspace = true }
 tokio = { workspace = true, features = ["full"] }
 tokio-stream = { version = "0.1.18", features = ["net"] }
@@ -40,7 +44,7 @@ dora-message = { workspace = true }
 serde_yaml = { workspace = true }
 uuid = { workspace = true }
 futures = { workspace = true }
-shared_memory_extended = "0.13.0"
+shared_memory_extended = { version = "0.13.0", optional = true }
 aligned-vec = "0.6.4"
 ctrlc = "3.5.2"
 which = "8"
